chore : 변수명 변경 + jpa 인자 변경
UserDto
- getCreateDate -> getCreateDateTime

UserRepository
- Integer -> Long

+++ src/main/java/com/zerobase/hoops/users/repository/UserRepository.java
@@ -1,13 +1,12 @@
 package com.zerobase.hoops.users.repository;
 
-import com.zerobase.hoops.entity.BlackListUserEntity;
 import com.zerobase.hoops.entity.UserEntity;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<UserEntity, Integer> {
+public interface UserRepository extends JpaRepository<UserEntity, Long> {
 
   boolean existsById(String id);
